Add Cart, CartProduct, Image, Gallery model. Update @GeneratedValue(strategy=IDENTIFY) to models Update config files. Change Long to Integer in Services

+++ trunk/onlinestore/src/main/java/com/onlinestore/service/CategoryProductService.java
@@ -9,9 +9,9 @@ public interface CategoryProductService {
 	
 	List<CategoryProduct> getCategoryProducts();
 	
-	CategoryProduct getCategoryProduct(Long id);
+	CategoryProduct getCategoryProduct(Integer id);
 	
 	void updateCategoryProduct(CategoryProduct categoryProduct);
 	
-	void deleteCategoryProduct(Long id);
+	void deleteCategoryProduct(Integer id);
 }
